restructure packages
