+++ internal/hive/plugins/claude/cache.go
