+++ internal/hive/plugins/claude/claude.go
@@ -12,8 +12,8 @@ import (
 	"charm.land/lipgloss/v2"
 	"github.com/hay-kot/hive/internal/core/config"
 	"github.com/hay-kot/hive/internal/core/session"
-	"github.com/hay-kot/hive/internal/plugins"
-	"github.com/hay-kot/hive/internal/styles"
+	"github.com/hay-kot/hive/internal/core/styles"
+	"github.com/hay-kot/hive/internal/hive/plugins"
 )
 
 // Plugin implements Claude Code integration (fork + analytics).
