+++ internal/hive/plugins/claude/claude_test.go
