+++ internal/hive/plugins/claude/parser.go
