+++ internal/hive/plugins/collector.go
