+++ internal/hive/plugins/manager.go
