+++ internal/hive/plugins/manager_test.go
