+++ internal/hive/plugins/pool.go
