+++ internal/hive/scripts/bin/agent-send
