+++ internal/hive/scripts/scripts.go
