+++ internal/hive/scripts/scripts_test.go
