+++ internal/tui/views/review/review_view.go
@@ -20,8 +20,8 @@ import (
 	"github.com/rs/zerolog/log"
 
 	corereview "github.com/hay-kot/hive/internal/core/review"
+	"github.com/hay-kot/hive/internal/core/styles"
 	"github.com/hay-kot/hive/internal/data/stores"
-	"github.com/hay-kot/hive/internal/styles"
 	"github.com/hay-kot/hive/internal/tui/components"
 )
