+++ internal/tui/views/review/review_view_test.go
@@ -11,9 +11,9 @@ import (
 
 	tea "charm.land/bubbletea/v2"
 
+	"github.com/hay-kot/hive/internal/core/styles"
 	"github.com/hay-kot/hive/internal/data/db"
 	"github.com/hay-kot/hive/internal/data/stores"
-	"github.com/hay-kot/hive/internal/styles"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
