+++ main.go
@@ -15,15 +15,15 @@ import (
 	"github.com/hay-kot/hive/internal/data/db"
 	"github.com/hay-kot/hive/internal/data/stores"
 	"github.com/hay-kot/hive/internal/hive"
-	"github.com/hay-kot/hive/internal/plugins"
-	"github.com/hay-kot/hive/internal/plugins/beads"
-	"github.com/hay-kot/hive/internal/plugins/claude"
-	"github.com/hay-kot/hive/internal/plugins/contextdir"
-	"github.com/hay-kot/hive/internal/plugins/github"
-	"github.com/hay-kot/hive/internal/plugins/lazygit"
-	"github.com/hay-kot/hive/internal/plugins/neovim"
-	plugintmux "github.com/hay-kot/hive/internal/plugins/tmux"
-	"github.com/hay-kot/hive/internal/scripts"
+	"github.com/hay-kot/hive/internal/hive/plugins"
+	"github.com/hay-kot/hive/internal/hive/plugins/beads"
+	"github.com/hay-kot/hive/internal/hive/plugins/claude"
+	"github.com/hay-kot/hive/internal/hive/plugins/contextdir"
+	"github.com/hay-kot/hive/internal/hive/plugins/github"
+	"github.com/hay-kot/hive/internal/hive/plugins/lazygit"
+	"github.com/hay-kot/hive/internal/hive/plugins/neovim"
+	plugintmux "github.com/hay-kot/hive/internal/hive/plugins/tmux"
+	"github.com/hay-kot/hive/internal/hive/scripts"
 	"github.com/hay-kot/hive/pkg/executil"
 	"github.com/hay-kot/hive/pkg/logutils"
 	"github.com/hay-kot/hive/pkg/tmpl"
